+++ inst/tinytest/test_diversity.R
@@ -3,6 +3,10 @@ Sys.setenv(LANGUAGE = "en") # Force locale
 source("helpers.R")
 data("cantabria")
 
+# Diversity ====================================================================
+d <- diversity(cantabria)
+expect_identical(dim(d), c(5L, 11L))
+
 # Heterogeneity ================================================================
 method <- c("berger", "brillouin", "mcintosh", "simpson", "shannon")
 for (i in method) {
@@ -11,7 +15,8 @@ for (i in method) {
   expect_equal(get_method(index), i)
 }
 
-boot <- with_seed(12345, bootstrap(index, n = 30))
+boot <- suppressWarnings(bootstrap(index, n = 30, seed = 12345))
+expect_true(all(boot$bias < 0)) # Downward bias
 expect_equal_to_reference(boot, file = "_snaps/heterogeneity_bootstrap.rds")
 
 jack <- jackknife(index)
@@ -25,7 +30,7 @@ for (i in method) {
   expect_equal(get_method(index), i)
 }
 
-boot <- with_seed(12345, bootstrap(index, n = 30))
+boot <- suppressWarnings(bootstrap(index, n = 30, seed = 12345))
 expect_equal_to_reference(boot, file = "_snaps/evenness_bootstrap.rds")
 
 jack <- jackknife(index)
@@ -71,13 +76,8 @@ expect_equal_to_reference(test_simpson(birds),
 
 # Plot =========================================================================
 if (at_home()) {
-  source("helpers.R")
   using("tinysnapshot")
-  options(tinysnapshot_device = "svglite")
-  options(tinysnapshot_height = 7) # inches
-  options(tinysnapshot_width = 7)
-  options(tinysnapshot_tol = 200) # pixels
-  options(tinysnapshot_os = "Linux")
+  source("helpers.R")
 
   idx_heterogeneity <- with_seed(12345, {
     idx_heterogeneity <- heterogeneity(cantabria, method = "shannon")